requirements.txt
fastapi[standard]
openai
jinja2
pyjwt
pwdlib[argon2]
sqlalchemy
alembic
python-dotenv